Add files via upload
In this project, we have to create a data model and Power BI report for Seven Sages Tea Company (SSTC) that combines information from all over the company. 
The data model will make it possible for the company's CFO to quickly review and analyze which tea beverages sell well and which ones generate the highest profitability.

Seven Sages Tea Company (SSTC) up near Seattle, Washington (USA) is a small, but growing company, and the CFO is eager to get some help translating all the disjointed sales files and product records into something usable so that they can help the owners make sense of what is generating not only sales but profitability.

It's common knowledge that the flagship tea beverage, Bamboo Grove Tea is the company's biggest seller. But what about the rest? The way the data is collected in-house, they are not even sure whether all the products are generating a profit at this stage.

Here are the main steps for the project. 

-Download and familiarize yourself with the source files provided by SSTC.

-Use Get Data to load the data from the starter materials into Power BI.

-Clean and format your data so that it will work well in your data model.

-Create a date table to support time intelligence.

-Build a relationship from each dimension to the relevant key on the fact table.

-To satisfy the CFO's requirements, we will write six measures—to calculate Sales, Cost of Sales and Gross Profit Margin in two different currencies.

-Build a basic visual report to display your findings.